backend/requirements.txt
torch>=2.0.0
diffusers>=0.20.0
transformers>=4.20.0
Pillow>=9.0.0
numpy>=1.21.0
